+++ src/main/java/com/group/mis_servicios/model/repository/CredentialsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CredentialsRepository extends JpaRepository<Credentials, Integer> {
+public interface CredentialsRepository extends JpaRepository<Credentials, Long> {
     Optional<Credentials> findByUsername(String username);
 }
