+++ src/main/java/com/group/mis_servicios/model/repository/FacilityRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface FacilityRepository extends JpaRepository<Facility, Integer> {
+public interface FacilityRepository extends JpaRepository<Facility, Long> {
     Optional<Facility> findByName(String name);
 }
