+++ src/main/java/com/group/mis_servicios/view/dto/FavoritesResponseDTO.java
@@ -8,9 +8,9 @@ import java.util.List;
 @Getter
 @Setter
 public class FavoritesResponseDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private LocalDateTime creationDate;
-    private Integer clientId;
+    private Long clientId;
     private List<ProviderResponseDTO> providers;
 }
